AlgoTrading/LSTM_strategy/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tensorflow>=2.10.0
keras>=2.10.0
yfinance>=0.2.0
ta>=0.10.0
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.10.0
python-dotenv>=0.19.0
joblib>=1.1.0
